Reorganize CLI bin folder
