+++ bin/commands/check.js
@@ -1,0 +1,46 @@
+import { echo, error } from '../lib/shell.js';
+import { resolve } from '../lib/helpers.js';
+
+export default async function check() {
+	const pkg = await loadPluginPackageInfo();
+
+	echo(`Checking package info of ${pkg.name}`);
+
+	const errors = [];
+
+	if (!pkg.amdName) {
+		errors.push('package.json missing `amdName` property');
+	}
+	if (pkg.type !== 'module') {
+		errors.push('package.json `type` property must be "module"');
+	}
+	if (!pkg.source) {
+		errors.push('package.json missing `source` property');
+	}
+	if (!pkg.main) {
+		errors.push('package.json missing `main` property');
+	}
+	if (!pkg.module) {
+		errors.push('package.json missing `module` property');
+	}
+	if (!pkg.unpkg) {
+		errors.push('package.json missing `unpkg` property');
+	}
+	if (!pkg.exports) {
+		errors.push('package.json missing `exports` property');
+	}
+
+	if (errors.length) {
+		error(errors);
+		return false;
+	} else {
+		echo('package.json looks good!');
+		return true;
+	}
+};
+
+async function loadPluginPackageInfo() {
+	const importPath = resolve('package.json');
+	const { default: pkg } = await import(importPath, { assert: { type: 'json' } });
+	return pkg;
+}
