+++ bin/lib/shell.js
@@ -1,26 +1,21 @@
 import shell from 'shelljs';
-
-export function cwd() {
-	return process.cwd();
-}
-
-export function pwd() {
-	return shell.pwd().toString();
-}
+import live from 'shelljs-live';
+import chalk from 'chalk';
 
 export function exec(command) {
-	return shell.exec(command);
+	return live(command);
+	// return shell.exec(command);
 }
 
 export function echo(message) {
-	shell.echo(`[swup] ${message}`);
+	console.log(chalk.gray('[swup]'), message);
 }
 
 export function error(message) {
 	if (Array.isArray(message)) {
 		message.forEach(msg => error(msg));
 	} else {
-		echo(`Error: ${message}`);
+		console.log(chalk.red.bold(message));
 	}
 	shell.exit(1);
 }
