--- cli/commands/lint.js
@@ -1,7 +1,0 @@
-import { exec, echo } from '../lib/shell.js';
-
-export default async function lint() {
-  echo('Linting plugin with prettier');
-  exec('prettier src/**/*.{js,mjs} --write');
-  return true;
-};
